+++ examples/resources/nxos_dhcp_relay_address/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_dhcp_relay_address.example "sys/dhcp/inst/relayif-[eth1/10]/addr-[VRF1]-[1.1.1.1]"