+++ internal/provider/provider.go
@@ -240,6 +240,7 @@ func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceT
 		"nxos_subinterface":           resourceSubinterfaceType{},
 		"nxos_subinterface_vrf":       resourceSubinterfaceVRFType{},
 		"nxos_dhcp_relay_interface":   resourceDHCPRelayInterfaceType{},
+		"nxos_dhcp_relay_address":     resourceDHCPRelayAddressType{},
 	}, nil
 }
 
@@ -265,6 +266,7 @@ func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSou
 		"nxos_subinterface":           dataSourceSubinterfaceType{},
 		"nxos_subinterface_vrf":       dataSourceSubinterfaceVRFType{},
 		"nxos_dhcp_relay_interface":   dataSourceDHCPRelayInterfaceType{},
+		"nxos_dhcp_relay_address":     dataSourceDHCPRelayAddressType{},
 	}, nil
 }