+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add nxos_subinterface resource and data source
 - Add nxos_subinterface_vrf resource and data source
 - Add nxos_dhcp_relay_interface resource and data source
+- Add nxos_dhcp_relay_address resource and data source
 
 ## 0.2.3